adds previous Workflow components into ConfigChange folder and renames all workflow components to ConfigChange
