+++ react-front/public/components/ConfigChange/ConfigChangeProgressBar.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+class ConfigChangeProgressBar extends React.Component {
+  render() {
+    return (
+      <div>
+        <p>This a footer</p>
+      </div>
+    );
+  }
+}
+
+export default ConfigChangeProgressBar;
