+++ react-front/public/components/ConfigChange/ConfigChangeStep4.js
@@ -1,0 +1,21 @@
+import React from "react";
+
+class ConfigChangeStep4 extends React.Component {
+  render() {
+    return (
+      <div className="workflow-container">
+        <div className="workflow-container__header">
+          <h2>Commit configuration (4/4)</h2>
+          <a href="#">
+            <button className="workflow-container__button--hide">Close</button>
+          </a>
+        </div>
+        <div className="workflow-collapsable">
+          <p>this is box4 </p>
+        </div>
+      </div>
+    );
+  }
+}
+
+export default ConfigChangeStep4;
